+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **[UPDATE]** Update `Itinerary` to handle segments and connections
 - **[NEW]** Add `ConnectionIcon`
+- **[UPDATE]** Add `subText` to `TripCard`'s driver
+  [...]
 
 # v40.7.0 (01/10/2020)